Refactored and worked on FeatureReader.
Renamed FeatureSection class to FeaturePart. FeatureSection is now
name of a FeatureFileSection sub-class. 'Parts' of a feature,
'sections' of a feature file.

Added a FakeFreemindBuilder class to spec_helper.rb but not
sure that is where it should go.

Started work on next steps for FeatureReader. Create a 'tmp'
directory for files created by tests.